backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
stripe==7.6.0
pillow==10.1.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0 